+++ src/Entities/Damage/IDamage.java
@@ -3,9 +3,9 @@ package Entities.Damage;
 public interface IDamage {
     public DamageType getDamageType();
 
-    public int getValue();
+    public Double getValue();
 
     public void setDamageType(DamageType damageType);
 
-    public void setValue(int value);
+    public void setValue(Double value);
 }
